+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Changed PluginConfig to be more consistent with new name
 - Changed auth to support instance roles, shared credentials
 - Added auth support for assuming a role
+- Added deprecation warnings about using key arguments
 
 ## [0.2.1] - 2020-02-10